Allow both static and dynamic linking with raylib
This commit includes raylib as a submodule for Mason, which by default
will be built when running make and the object files from raylib and
Mason will be archived into a single static library. If the user instead
wishes to link with raylib dynamically, the environment variables
RAYLIB_LIBTYPE and RAYLIB_INCLUDE_DIR must be set to DYNAMIC and to the
path which includes the headers for raylib respectivelly, and the raylib
submodule does not need to be cloned.

+++ makefile
@@ -2,7 +2,10 @@ CC=clang
 CFLAGS ?= -Wall -Wextra -Werror -Wno-unused-parameter -Wno-unused-but-set-variable -O2
 DEBUG=-fsanitize=address,undefined -g3
 
-INCLUDE=-Iinclude/ -IScaffold/include/
+RAYLIB_LIBTYPE ?= STATIC
+RAYLIB_INCLUDE_DIR ?= raylib/src/
+
+INCLUDE=-Iinclude/ -IScaffold/include/ -I${RAYLIB_INCLUDE_DIR}
 
 SRC_DIR=src
 OBJ_DIR=build
@@ -12,11 +15,19 @@ SOURCES=$(wildcard ${SRC_DIR}/*.c)
 
 OUTPUT=libmason.a
 
-final: compile move assemble clean
+.PHONY: final debug enable-debug raylib compile move assemble clean
+
+final: raylib compile move assemble
 debug: enable-debug final
 
 enable-debug:
 	$(eval CFLAGS := ${CFLAGS} ${DEBUG})
+
+raylib:
+ifeq (${RAYLIB_LIBTYPE},STATIC)
+	@echo [building raylib...]
+	$(MAKE) -C raylib/src/
+endif
 
 compile:
 	@echo [compiling...]
@@ -37,9 +48,18 @@ assemble:
 	#create necessary dir if it doesn't exist
 	test -d ${OUT_DIR} || mkdir ${OUT_DIR}
 
-	# assemble object files into final static library file
-	ar rs ${OUT_DIR}/${OUTPUT} $(wildcard ${OBJ_DIR}/*.o)
+	# assemble object files and raylib into final static library file
+	$(eval objs := $(wildcard ${OBJ_DIR}/*.o))
+ifeq (${RAYLIB_LIBTYPE},STATIC)
+	$(eval objs += $(wildcard ${RAYLIB_INCLUDE_DIR}/*.o))
+endif
+	ar -rcs ${OUT_DIR}/${OUTPUT} ${objs}
 
 clean:
+ifneq ($(wildcard ${OBJ_DIR}/*.o),)
 	rm $(wildcard ${OBJ_DIR}/*.o)
+endif
+ifeq (${RAYLIB_LIBTYPE}, STATIC)
+	$(MAKE) clean -C raylib/src/
+endif
